separated team api from auth api. Implemented team member deactivation & delete endpoints. Added checkUserAdminMiddleware
