+++ server/src/app.js
@@ -4,6 +4,7 @@ import cors from 'cors';
 import authRoutes from './routes/auth.route.js';
 import taskRoutes from './routes/task.route.js';
 import projectRoutes from "./routes/project.route.js"
+import teamRoutes from "./routes/team.route.js"
 
 dotenv.config();
 
@@ -15,6 +16,7 @@ app.use(express.json());
 
 // Routes
 app.use('/api/auth', authRoutes);
+app.use('/api/team', teamRoutes);
 app.use('/api/tasks', taskRoutes);
 app.use('/api/projects', projectRoutes);
